feat: add .gitignore, index.js, constants.js. Install babel
