+++ extra/patterns/index.js
@@ -1,0 +1,2 @@
+export { default as diamond }from './diamond';
+export { default as equilateral }from './equilateral';
