+++ extra/utils/index.js
@@ -1,0 +1,4 @@
+export { default as hasPermissions }from './permissions';
+export { default as validateUser }from './validation';
+
+
