+++ extra/utils/permissions.js
@@ -1,15 +1,9 @@
-let permissions = {
-    'getUsers': {
-        all: ['head-trainer'],
-        read:['trainee', 'trainer'],
-        write: ['trainer'],delete: []
-    }
-};
-function hasPermissions(module, role, permissionType)  {
+import {permissions} from './../constants';
+export default function hasPermissions(module, role, permissionType)  {
     let k = permissions[module];
     let b = k[permissionType];
     return b.some(element =>{
         return element == role;
     });
 }
-console.log(hasPermissions('getUsers', 'trainer', 'write'));
+//console.log(hasPermissions('getUsers', 'trainer', 'write'));
